Allow mouse deltas to be 32bit ints

+++ lib/ble/ble.cpp
@@ -305,7 +305,7 @@ void decryptSendString(SecureSession::rawDataPacket* packet, SecureSession* sess
     
     // The first byte indicates the data is mouse data
     else if(plaintext[0] == 2){
-      std::vector<int16_t> keycode(plaintext + 1, plaintext + packet->dataLen);
+      std::vector<int32_t> keycode(plaintext + 1, plaintext + packet->dataLen);
       moveMouse(keycode.data());
     }
